+++ src/Lib/MceObject.php
@@ -301,7 +301,7 @@ abstract class MceObject implements Serializable {
 	 *          Liste des champs pour lesquels on ne sera pas sensible à la casse
 	 * @return Contact[] Array
 	 */
-	function getList($fields = [], $filter = "", $operators = [], $orderby = "", $asc = true, $limit = null, $offset = null, $case_unsensitive_fields = []) {
+	public function getList($fields = [], $filter = "", $operators = [], $orderby = "", $asc = true, $limit = null, $offset = null, $case_unsensitive_fields = []) {
 		M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->getList()");
 		$_objects = $this->objectmelanie->getList($fields, $filter, $operators, $orderby, $asc, $limit, $offset, $case_unsensitive_fields);
 		if (!isset($_objects)) {